Add autocap file
